Refactor: Define PodMetricsClient interface and hide implementation details of vllm metrics processing
